Run robot operations in seperate threads
We wish to perform checks that operations start and finish correctly
on the server rather than the client. This requires the operation initiator
functions run for a significant length of time. However the request
handler function needs to respond quickly to prevent blocking other
incoming requests. To handle this we start the operation function in a
new thread. We add a decorator to prevent more than one foreground
operation being started at a time.