Enable clients to query the server
Clients may need to do direct queries of the server state. These are
short running operations that don't need to be broadcast to all clients.
This change adds a new 'query' operation type to enable these queries.
We also add integration tests to check the client and server are
interacting correctly.