Require foreground done for foreground operations
When robot foreground operations are started directly through EPICS
users should not be able to start a foreground operation through the
server. This change makes the foreground_operation decorator check the
foreground_done PV before running the operation.

This change also declares several Server attributes and methods private
by adding a leading underscore.

We remove the numpy dependency as it is not being used and takes
significant time to install.

We have dropped support for Python 2 and so no longer need the six or mock
packages.