Add current_task attribute to robot
A current task PV has been added to RobotEpsonIP. This change exposes the
variable as the current_task attribute on the Robot and RobotClient
classes.

We also rename run_args to task_args for clarity and
run_foreground_operation to run_task to better distinguish robot tasks
from server operations.